SwapFyterzMain cleanup. Code organization changes
