+++ core/src/com/dev/swapftrz/menu/SPFZMenuAction.java
@@ -311,4 +311,6 @@ public class SPFZMenuAction
   //STAGE SELECT
   public void processStageSelect() {
   }
+
+
 }
